feat: React Portal 기반 모달 프로바이더 구현
- zustand를 사용한 모달 상태 관리 추가
- React Portal을 활용한 ModalProvider 컴포넌트 구현
- layout.tsx에 ModalProvider 연결하여 children 감싸기
- max-w-md, w-full 제거한 모달 wrapper 스타일 적용
- 배경 클릭 시 모달 닫기 기능 구현
