+++ jiyoontrip/src/app/commons/providers/modal/modal.provider.module.css
@@ -1,17 +1,18 @@
 .backdrop {
+  display: flex;
   position: fixed;
   top: 0;
   left: 0;
   width: 100%;
   height: 100%;
   background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
   justify-content: center;
   align-items: center;
   z-index: 1000;
 }
 
-.wrapper {
+.modal {
+  display: flex;
   background-color: white;
   border-radius: 8px;
   box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
